Moving TestRunner into separate file
